requirements.txt
numpy>=2.1.2
pandas>=2.2.3
scikit-learn>=0.24.0
scipy>=1.7.0
statsmodels>=0.13.0
yfinance>=0.1.63
ta>=0.11.0
torch>=1.9.0
torch-directml>=0.1.0
optuna>=2.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
pyyaml>=5.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.5b2
pylint>=2.8.2
mypy>=0.910
pandas_datareader>=0.10.0
TA-Lib>=0.4.32
xgboost>=1.5.0
lightgbm>=3.3.1
markdown2>=2.4.0
pdfkit>=0.6.1
plotly>=5.0.0
cvxopt>=1.2.7
pathlib>=1.0.1
